Prevent creation of new peers when decommissioning of a node was working
Decommissioning and leav event could tear down ClientNodes already, so ensure to not create another one when handling the rest of the decommissioning

Addresses https://github.com/matter-js/matterjs-server/issues/45

+++ packages/matter.js/src/MatterController.ts
@@ -520,7 +520,11 @@ export class MatterController {
 
     async removeNode(nodeId: NodeId) {
         const peerAddress = this.fabric.addressOf(nodeId);
-        const node = await this.node.peers.forAddress(peerAddress);
+        const node = this.node.peers.get(peerAddress);
+        if (!node) {
+            // Unknown or already removed
+            return;
+        }
         const peer = this.node.env.get(PeerSet).for(peerAddress);
         await node.delete();
         await peer.delete();
